feat: ensure ChangeProfilePictureService not calls DeleteFile if file does not exists and SaveUserPictureRepo throws

+++ src/domain/services/change-profile-picture.ts
@@ -33,7 +33,10 @@ export class ChangeProfilePictureService {
     try {
       await this.userProfileRepo.savePicture(userProfile);
     } catch {
-      await this.fileStorage.delete({ key });
+      if (file) {
+        await this.fileStorage.delete({ key });
+      }
+      throw new Error();
     }
 
     return {
